forgot to allow the embeds to not preserve the avatar in TEST guild

+++ wall_e/utilities/embed.py
@@ -117,24 +117,27 @@ async def embed(logger, ctx=None, interaction=None, title='', content=None, desc
     emb_obj = discord.Embed(title=title, type='rich')
     emb_obj.description = description
     if avatar_url != "":
-        from wall_e_models.models import EmbedAvatar
-        # the below is needed in case the avatar url that was passed in to this function is deleted at some point
-        # in the future, which will result in an embed that has a broken avatar
-        avatar_obj = await EmbedAvatar.get_avatar_by_url(avatar_url)
-        if avatar_obj is None:
-            channels = interaction.guild.channels if interaction is not None else ctx.guild.channels
-            embed_avatar_chan: discord.TextChannel = discord.utils.get(channels, name="embed_avatars")
-            avatar_file_name = f'avatar-{time.time()*1000}.png'
-            with open(avatar_file_name, "wb") as file:
-                file.write(requests.get(avatar_url).content)
-            avatar_msg = await embed_avatar_chan.send(file=discord.File(avatar_file_name))
-            os.remove(avatar_file_name)
-            avatar_obj = EmbedAvatar(
-                avatar_discord_url=avatar_url,
-                avatar_discord_permanent_url=avatar_msg.attachments[0].url
-            )
-            await EmbedAvatar.insert_record(avatar_obj)
-        avatar_url = avatar_obj.avatar_discord_permanent_url
+        channels = interaction.guild.channels if interaction is not None else ctx.guild.channels
+        embed_avatar_chan: discord.TextChannel = discord.utils.get(channels, name="embed_avatars")
+        if embed_avatar_chan is not None:  # since the TEST guild is the only environment where the channel does
+            # not exist this is currently the best way to ensure that embed avatars aren't preserved for longevity
+            # on the TEST guils since there's really no point to that on a TEST environment
+            from wall_e_models.models import EmbedAvatar
+            # the below is needed in case the avatar url that was passed in to this function is deleted at some point
+            # in the future, which will result in an embed that has a broken avatar
+            avatar_obj = await EmbedAvatar.get_avatar_by_url(avatar_url)
+            if avatar_obj is None:
+                avatar_file_name = f'avatar-{time.time()*1000}.png'
+                with open(avatar_file_name, "wb") as file:
+                    file.write(requests.get(avatar_url).content)
+                avatar_msg = await embed_avatar_chan.send(file=discord.File(avatar_file_name))
+                os.remove(avatar_file_name)
+                avatar_obj = EmbedAvatar(
+                    avatar_discord_url=avatar_url,
+                    avatar_discord_permanent_url=avatar_msg.attachments[0].url
+                )
+                await EmbedAvatar.insert_record(avatar_obj)
+            avatar_url = avatar_obj.avatar_discord_permanent_url
     emb_obj.set_author(name=author, icon_url=avatar_url)
     emb_obj.colour = COLOUR_MAPPING[colour]
     emb_obj.set_thumbnail(url=thumbnail)
